+++ app/src/js/components/Help/getting_started_data.js
@@ -1,0 +1,110 @@
+const GettingStartedData = {
+    "id":5,
+    "heading":"<h1>Getting Started<hr></h1>",
+    "paragraphs":[
+      {
+         "text":"The steps below will guide you through how to use Earthdata Pub to publish your data with a <a title='List of NASA DAACs' href='data_publication_guidelines#daacs'>NASA DAAC</a>."
+      },
+       {
+          "text":"For more information on the data publication process, see the <a title='data publication process' href='data_publication_guidelines#publication'>Data Publication Process</a> section of the <a href='data_publication_guidelines' name='Data Publication Guidelines' link_title='Data Publication Guidelines' link_text='Data Publication Guidelines'>Data Publication Guidelines</a>."
+       },
+       {
+         "heading":"<h2 class='display-4' id='account'>Create an Earthdata Pub Account<hr></h2>",
+         "text":"To create an Earthdata Pub account, please <a href='javascript:feedback.showForm();' title='Send feedback using feedback form.'>contact us</a>. If you know at which <a title='List of NASA DAACs' href='data_publication_guidelines#daacs'>DAAC</a> you would like to publish your data, please include that in your request. "
+       },
+       {
+         "text": `When your account has been created, you will receive an email with your username and temporary password. You then need to log in and change your password, as the temporary passwords will expire. When you first log in, you will receive instructions on setting up Multi-Factor Authentication (MFA) <style>.mfa-tooltip { position: relative; display: inline-block;}.mfa-tooltip .mfa-tooltiptext { visibility: hidden; width: 500px; background-color: #555; color: #fff; text-align: center; border-radius: 6px; padding: 5px 0; position: absolute; z-index: 1; top: 125%; left: 50%; margin-left: -250px; opacity: 0; transition: opacity 0.3s;}.mfa-tooltip .mfa-tooltiptext::after { content: ""; position: absolute; bottom: 100%; left: 50%; margin-left: -5px; border-width: 5px; border-style: solid; border-color: transparent transparent #555 transparent;}.mfa-tooltip:hover .mfa-tooltiptext { visibility: visible; opacity: 1;}</style><div class="mfa-tooltip"><svg aria-hidden="true" focusable="false" data-prefix="fas" data-icon="circle-info" class="svg-inline--fa fa-circle-info " role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path fill="#2276ac" d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM216 336l24 0 0-64-24 0c-13.3 0-24-10.7-24-24s10.7-24 24-24l48 0c13.3 0 24 10.7 24 24l0 88 8 0c13.3 0 24 10.7 24 24s-10.7 24-24 24l-80 0c-13.3 0-24-10.7-24-24s10.7-24 24-24zm40-208a32 32 0 1 1 0 64 32 32 0 1 1 0-64z"></path></svg><span class="mfa-tooltiptext">Earthdata Pub uses MFA, a security system that requires users to provide two or more verification factors to gain access to an account or system. These factors typically include something you know (like a password) and something you have (like a smartphone or security token). MFA is important because it adds an extra layer of protection, significantly reducing the risk of unauthorized access even if one factor, such as a password, is compromised.Please contact us if you have problems setting up or using MFA.</span></div>`
+      },
+       {
+          "heading":"<h2 class='display-4' id='accession'>Data Accession Request<hr></h2>",
+          "text":"If your data have not been assigned to a DAAC, you will need to have your data approved through the Data Accession process. For more information on this process, see the <a title='Data Accession Process' href='/data_publication_guidelines#accession'>Data Accession Process</a> section of the <a href='data_publication_guidelines' name='Data Publication Guidelines' link_title='Data Publication Guidelines' link_text='Data Publication Guidelines'>Data Publication Guidelines</a>."
+       },
+       {
+          "step":[
+             {
+                "number":1,
+                "heading":"<span class='main-width sections'><b>Start a new Data Accession Request</b></span>",
+                "text":"Log into the <a title='Earthdata Pub Dashboard' href='/dashboard'>Earthdata Pub Dashboard</a> with your Earthdata Pub account. To create a Data Accession request, click on “New Request”."
+             },
+             {
+                "number":2,
+                "heading":"<span class='main-width sections'><b>Fill in the Data Accession Request form</b></span>",
+                "text":"In the <a title='Earthdata Pub Dashboard' href='/dashboard'>Earthdata Pub Dashboard</a>, click on the “Data Accession Request Form” button in the Next Action column. The information you provide about your data will be used to review and approve your Data Accession Request. ",
+                "icon":"lightbulb.svg",
+                "icon_text":"You can save your progress and return to the form later."
+             },
+             {
+                "number":3,
+                "heading":"<span class='main-width sections'><b>Submit your request</b></span>",
+                "text":"After you have filled in the Data Accession Request form, click on “Submit” to submit the request. This will initiate the Data Accession process. For more information on this process, see the <a title='Data Accession Process' href='/data_publication_guidelines#accession'>Data Accession Process</a> section of the <a href='data_publication_guidelines' name='Data Publication Guidelines' link_title='Data Publication Guidelines' link_text='Data Publication Guidelines'>Data Publication Guidelines</a>."
+             },
+             {
+                "number":4,
+                "heading":"<span class='main-width sections'><b>Accession request is completed</b></span>",
+                "paragraphs": [
+                  {
+                     "text":"Your data will be evaluated by NASA Earth Science Data and Information Systems personnel who will determine whether your data are approved to be published at a DAAC. The decision will be documented in the Earthdata Pub Dashboard, and you will receive an email explaining the decision."
+                  },
+                  {
+                     "text":"If your Data Accession request has been approved, you can move to the data publication process."
+                  }
+               ]
+            }
+          ]
+       },
+       {
+          "heading":"<h2 class='display-4' id='publication'>Data Publication Request<hr></h2>",
+          "text":"Once your data have been assigned to a DAAC, you can complete a Data Publication request form when you are ready to work with the DAAC to publish your data product. For more information on the data publication process, see the <a title='data publication process' href='data_publication_guidelines#publication'>Data Publication Process</a> section of the <a href='data_publication_guidelines' name='Data Publication Guidelines' link_title='Data Publication Guidelines' link_text='Data Publication Guidelines'>Data Publication Guidelines</a>."
+       },
+       {
+          "step":[
+             {
+                "number":1,
+                "heading":"<span class='main-width sections'><b>Fill in the Data Publication Request form</b></span>",
+                 "text":"In the <a title='Earthdata Pub Dashboard' href='dashboard/'>Earthdata Pub Dashboard</a>, click on the “Data Publication Request Form” button in the Next Action column. The form will be pre-populated with the information you provided in the Data Accession form, if you completed that form.",
+                "icon":"lightbulb.svg",
+                "icon_text":"You can save your progress and return to the form later."
+             },
+             {
+                "number":2,
+                "heading":"<span class='main-width sections'><b>Submit your request</b></span>",
+                "text":"After you have filled in the Data Publication request form, click on “Submit” to submit the request. This will initiate the data publication process."
+             },
+             {
+                "number":3,
+                "heading":"<span class='main-width sections'><b>Collaborate with the DAAC</b></span>",
+                "paragraphs": [
+                  {
+                     "text":"The data publication process is a collaboration between you and the DAAC. The information you provide in the Data Publication form, along with sample data and related documentation, will be reviewed by the DAAC. You may be asked to provide additional information and will receive further instructions for submitting your entire data product."
+                  },
+                  {
+                      "text":"For more information on the typical activities performed during data publication, as well as your role and the role of the DAAC, see the <a title='data publication process' href='data_publication_guidelines#publication'>Data Publication Process</a> section of the <a href='data_publication_guidelines' name='Data Publication Guidelines' link_title='Data Publication Guidelines' link_text='Data Publication Guidelines'>Data Publication Guidelines</a>."
+                  },
+                  {
+                     "icon":"sticky-note.svg",
+                     "icon_text":"Some data publication steps and services may vary between DAACs."
+                  }
+                ]
+             },
+             {
+               "number":4,
+               "heading":"<span class='main-width sections'><b>Data product is published</b></span>",
+                "text":"When your data product has been published, you will receive an email from the DAAC and your request status will be updated in the <a title='Earthdata Pub Dashboard' href='dashboard'>Earthdata Pub Dashboard</a>."
+            }
+          ]
+       },
+       {
+          "heading":"<h2 class='display-4' id='status'>Track Your Request Status<hr></h2>",
+           "text":"You can track the status of your Data Accession or Data Publication requests on the Requests page of the <a title='Earthdata Pub Dashboard' href='dashboard'>Earthdata Pub Dashboard</a>. The Status column provides the current status of your request. Clicking on the status will provide additional information. "
+       },
+       {
+          "text":"You can also filter your requests by status by using the links on the left side of the Requests page. "
+       },
+       {
+         "heading":"<h2 class='display-4' id='communicate'>Communicate with the DAAC<hr></h2>",
+          "text":"You can communicate about your data product using the Conversations page of the <a title='Earthdata Pub Dashboard' href='dashboard'>Earthdata Pub Dashboard</a>. Communication can be conducted on the Conversations page during the Data Accession or Data Publication process."
+      }
+    ]
+};
+
+export default GettingStartedData;