+++ app/src/js/utils/privileges.js
@@ -111,7 +111,7 @@ export const requestCanReview = (privileges, stepName) => {
     return privileges.REQUEST.includes("REVIEW_MANAGER");
   }
 
-  return privileges.REQUEST.includes("REVIEW");
+  return privileges.REQUEST.includes("REVIEW") || privileges.REQUEST.includes("REVIEW_MANAGER");
 }
 
 export const formPrivileges = (privileges) => {